mujoco_simulator_py/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]

[project]
name = "mujoco_simulator_py"
description = "Python interface to https://github.com/JafarAbdi/mujoco_simulator"
version = "0.0.1"
classifiers = ["Programming Language :: Python :: 3.10"]
requires-python = "==3.11"
dependencies = [
  "dm-control",
  "mujoco",
  "eclipse-zenoh>=1.0.2,<2",
  "transforms3d>=0.4.2,<0.5",
  "loop-rate-limiters",
  "rich",
  "numpy",
  "scipy",
]

[project.urls]
repository = "https://github.com/JafarAbdi/mujoco_simulator_py"

[tool.setuptools.packages.find]
include = ["mujoco_simulator_py"]

[project.scripts]

[tool.mypy]
python_version = '3.10'
no_implicit_optional = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
